completed fixing the routes and middlewares
