+++ src/types/strapiTypes.ts
@@ -1,5 +1,3 @@
-import Stripe from 'stripe';
-
 export interface CustomerInterface {
     email: string;
 }
@@ -7,9 +5,4 @@ export interface CustomerInterface {
 export interface ProductInfoInterface {
     id?: string;
     name: string;
-}
-
-export interface PriceInfoInterface {
-    id?: string;
-    product: string  | Stripe.Product | Stripe.DeletedProduct;
 }
